requirements.txt
fastapi[standard]
sqlalchemy
psycopg
psycopg-binary
python-dotenv
